% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AccumulateNormal.R
\name{AccumulateNormal}
\alias{AccumulateNormal}
\title{SubtypeDrug internal function}
\usage{
AccumulateNormal(x_matrix, control_index)
}
\arguments{
\item{x_matrix}{A subpathway activity profile. rows are subpathwyas, columns are samples.}

\item{control_index}{A vector. In the sample of the subpathway activity profile, the position of control samples.}
}
\value{
A matrix.
}
\description{
Infering patient-specific subpathway activity profiles.
}
\details{
AccumulateNormal
}
\examples{
x<-matrix(c(1:10),ncol = 5)
x1<-AccumulateNormal(x,c(3,5))
}
\author{
Xudong Han,
Junwei Han,
Chonghui Liu
}
