% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CalculateSES.R
\name{CalculateSES}
\alias{CalculateSES}
\title{SubtypeDrug internal function}
\usage{
CalculateSES(labels.list, correl.vector = NULL)
}
\arguments{
\item{labels.list}{A vector of 0 and 1.}

\item{correl.vector}{A vector. The weight value used to calculate the enrichment score.}
}
\value{
A vector.
}
\description{
Calculate subpathway enrichment score.
}
\details{
CalculateSES
}
\examples{
x<-CalculateSES(sample(c(0,1),10,replace = TRUE),c(1:10))
}
\author{
Xudong Han,
Junwei Han,
Chonghui Liu
}
