% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Colork.R
\docType{data}
\name{Colork}
\alias{Colork}
\title{Color}
\format{
A vector containing 73 values.
}
\usage{
Colork
}
\description{
This variable stores the color data required by the program.
}
\examples{
data(Colork)

}
\keyword{datasets}
