% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Disease_drugs.R
\docType{data}
\name{Disease_drugs}
\alias{Disease_drugs}
\title{Simulated result data}
\format{
A list containing 8 variables.
The variables are as follows:
\itemize{
	\item Cacner Results table for cacner
	\item SubpathwayMatrix Subpathway activity natrix
	\item SampleInformation Cancer sample phenotypic information
	\item Parameter Parameter of the function PrioSubtypeDrug
}
}
\usage{
Disease_drugs
}
\description{
The simulated result data of only two sample types is generated by the functional PrioSubtypeDrug.
}
\examples{
# data(Disease_drugs)
}
\keyword{datasets}
