% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Drugs_CID.R
\docType{data}
\name{Drugs_CID}
\alias{Drugs_CID}
\title{Correspondence table of drug label and drug ID in PubCham database}
\format{
A dataframe with drug label and CID.
The variables are as follows:
\itemize{
	\item Drugs Drug label
	\item CID Drug ID in PubCham database
}
}
\usage{
Drugs_CID
}
\description{
A data frame for the drug and its corresponding PubCham database ID.
}
\examples{
data(Drugs_CID)
}
\keyword{datasets}
