% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeneexpT.R
\docType{data}
\name{GeneexpT}
\alias{GeneexpT}
\title{Gene expression data for testing}
\format{
A matrix with 40 samples.
}
\usage{
GeneexpT
}
\description{
Simulated normalized gene expression profile data.
}
\examples{
data(GeneexpT)
}
\keyword{datasets}
