% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDrugStructure.R
\name{getDrugStructure}
\alias{getDrugStructure}
\title{Get drug chemical structure diagram data}
\usage{
getDrugStructure(drug.label = "", main = "", sub = "")
}
\arguments{
\item{drug.label}{A character string of drug label to determine which drug to use for visualization.}

\item{main}{An overall title for the chemical structure graph.}

\item{sub}{A sub title for the chemical structure graph.}
}
\value{
A sdfset object.
}
\description{
`getDrugStructure()` outputs the chemical structure graph data of the
drug or compound based on the input drug label by the user. The results can be visualized by the `plot` function.
}
\details{
getDrugStructure
}
\examples{
require(rvest)
require(ChemmineR)
# Plot the chemical structure of drug pirenperone.
# Chem_str<-getDrugStructure(drug.label="pirenperone.")
# plot(Chem_str)
}
\author{
Xudong Han,
Junwei Han,
Chonghui Liu
}
