% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadClsFile.R
\name{ReadClsFile}
\alias{ReadClsFile}
\title{SubtypeDrug internal function}
\usage{
ReadClsFile(file)
}
\arguments{
\item{file}{Input sample subtype class vector file in CLS format.}
}
\value{
a list
}
\description{
These are function read sample label file (.cls format).
}
\details{
ReadClsFile
}
\examples{
Subtype<-system.file("extdata", "Subtype_labels.cls", package = "SubtypeDrug")
x<-ReadClsFile(Subtype)
}
\author{
Xudong Han,
Junwei Han,
Chonghui Liu
}
