% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Subtype_drugs.R
\docType{data}
\name{Subtype_drugs}
\alias{Subtype_drugs}
\title{Simulation result data}
\format{
A list containing 8 variables.
The variables are as follows:
\itemize{
	\item Basal Results table for basal subtype
	\item Her2 Results table for Her2 subtype
	\item LumA Results table for LumA subtype
	\item LumB Results table for LumB subtype
	\item DrugMatrix Drug disease reverse association matrix
	\item SubpathwayMatrix Subpathway activity natrix
	\item SampleInformation Cancer sample phenotypic information
	\item Parameter Parameter of the function OCSSD
}
}
\usage{
Subtype_drugs
}
\description{
The result data of the simulation is generated by the functional OCSSD.
}
\examples{
# data(Subtype_drugs)
}
\keyword{datasets}
