% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserDS.R
\docType{data}
\name{UserDS}
\alias{UserDS}
\title{Simulated user-defined drug regulator subpathway dataset}
\format{
A list containing 5 drugs.
}
\usage{
UserDS
}
\description{
The drug regulator subpathway data set is modeled as a case.
}
\examples{
data(UserDS)
}
\keyword{datasets}
