% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserDST.R
\docType{data}
\name{UserDST}
\alias{UserDST}
\title{User-defined drug regulator subpathway dataset for testing}
\format{
A list.
}
\usage{
UserDST
}
\description{
The drug regulator subpathway data set is modeled as a case.
}
\examples{
data(UserDST)
}
\keyword{datasets}
