% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserGS.R
\docType{data}
\name{UserGS}
\alias{UserGS}
\title{Simulated user-defined gene set data}
\format{
A list containing 50 gene sets.
}
\usage{
UserGS
}
\description{
Gene set data is simulated for case studies.
}
\examples{
data(UserDS)
}
\keyword{datasets}
