% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UserGST.R
\docType{data}
\name{UserGST}
\alias{UserGST}
\title{User-defined gene set data for testing}
\format{
A list.
}
\usage{
UserGST
}
\description{
Gene set data is simulated for case studies.
}
\examples{
data(UserGST)
}
\keyword{datasets}
